Use `elsa` frozen types in `ConstantPool` and `ClassBuilder`
Since these are append-only, using frozen datatypes makes it possible to
safely append new entries without needing exclusive access. The net
outcome is fewer `Rc<RefCell<..>>` types switching several `&mut` to `&`
simplifications.

OTOH, this does mean that `ConstantPool` has to re-implement a small
amount of the `OffsetVec` logic.

+++ lib/src/jvm/errors.rs
@@ -7,7 +7,7 @@ use super::{
 pub enum Error {
     ConstantPoolOverflow {
         constant: Constant,
-        offset: usize,
+        offset: u16,
     },
     IoError(std::io::Error),
     MethodCodeMaxStackOverflow(Offset),
